+++ src/App.vue
@@ -1,3 +1,7 @@
 <template id="app">
-    <div/>
+    <i class="fa fa-user text-info"/>
 </template>
+
+<script>
+export default { name: "App" };
+</script>