+++ CHANGELOG.md
@@ -26,3 +26,19 @@
 ### 📦 Packages
 
 * Set of required packages installed for Vue, lint and tests.
+* `@chenfengyuan/vue-countdown` installed with version `1.1.5`.
+* `@fortawesome/fontawesome-free` installed with version `5.15.4`.
+* `@popperjs/core` installed with version `2.10.2`.
+* `@sentry/vue` installed with version `6.15.0`.
+* `@sentry/tracing` installed with version `6.15.0`.
+* `axios` installed with version `0.24.0`.
+* `bootstrap` installed with version `5.1.3`.
+* `epic-spinners` installed with version `1.1.0`.
+* `qs` installed with version `6.10.1`.
+* `sass-loader` installed with version `12.3.0`.
+* `sass` installed with version `1.43.4`.
+* `sweetalert2` installed with version `11.2.0`.
+* `vue-i18n` installed with version `8.26.7`.
+* `vue-js-toggle-button` installed with version `1.3.3`.
+* `vue-toastification` installed with version `2.0.0-rc.5`.
+* `xss` installed with version `1.0.10`.